requirements.txt
requests
infisicalsdk
python-dotenv
kopf
kubernetes